Split calculator lib to specific directory
Split calculator libraries into net.nhiroki.lib.bluelinecalculator.

Still depends on androidx.annotation for Null treatment, but these are
just for compilers. For now keeping them and postponing spliting
completely removing androidx dependency.
