--- app/src/main/java/net/nhiroki/bluelineconsole/commands/calculator/FormulaPart.java
@@ -1,5 +1,0 @@
-package net.nhiroki.bluelineconsole.commands.calculator;
-
-public interface FormulaPart {
-
-}
