+++ app/src/main/java/net/nhiroki/lib/bluelinecalculator/CalculatorNumber.java
@@ -1,10 +1,10 @@
-package net.nhiroki.bluelineconsole.commands.calculator;
+package net.nhiroki.lib.bluelinecalculator;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import net.nhiroki.bluelineconsole.commands.calculator.units.CombinedUnit;
-import net.nhiroki.bluelineconsole.commands.calculator.units.UnitDirectory;
+import net.nhiroki.lib.bluelinecalculator.units.CombinedUnit;
+import net.nhiroki.lib.bluelinecalculator.units.UnitDirectory;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
