+++ app/src/main/java/net/nhiroki/lib/bluelinecalculator/FormulaPart.java
@@ -1,0 +1,5 @@
+package net.nhiroki.lib.bluelinecalculator;
+
+public interface FormulaPart {
+
+}
