+++ app/src/main/java/net/nhiroki/lib/bluelinecalculator/units/CombinedUnit.java
@@ -1,11 +1,11 @@
-package net.nhiroki.bluelineconsole.commands.calculator.units;
+package net.nhiroki.lib.bluelinecalculator.units;
 
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import net.nhiroki.bluelineconsole.commands.calculator.CalculatorExceptions;
-import net.nhiroki.bluelineconsole.commands.calculator.CalculatorNumber;
+import net.nhiroki.lib.bluelinecalculator.CalculatorExceptions;
+import net.nhiroki.lib.bluelinecalculator.CalculatorNumber;
 
 import java.util.ArrayList;
 import java.util.Arrays;
